+++ stroom-query/stroom-query-language/src/main/java/stroom/query/language/functions/ExpressionContext.java
@@ -26,21 +26,21 @@ public class ExpressionContext {
 
     private final int maxStringLength;
     private final DateTimeSettings dateTimeSettings;
-    private final StateFetcher stateFetcher;
+    private final StateProvider stateProvider;
 
     public ExpressionContext() {
         this.maxStringLength = 100;
         this.dateTimeSettings = DateTimeSettings.builder().build();
-        this.stateFetcher = (map, key, effectiveTimeMs) -> ValNull.INSTANCE;
+        this.stateProvider = (map, key, effectiveTimeMs) -> ValNull.INSTANCE;
     }
 
     @JsonCreator
     public ExpressionContext(final int maxStringLength,
                              final DateTimeSettings dateTimeSettings,
-                             final StateFetcher stateFetcher) {
+                             final StateProvider stateProvider) {
         this.maxStringLength = maxStringLength;
         this.dateTimeSettings = dateTimeSettings;
-        this.stateFetcher = stateFetcher;
+        this.stateProvider = stateProvider;
     }
 
     public int getMaxStringLength() {
@@ -51,8 +51,8 @@ public class ExpressionContext {
         return dateTimeSettings;
     }
 
-    public StateFetcher getStateFetcher() {
-        return stateFetcher;
+    public StateProvider getStateProvider() {
+        return stateProvider;
     }
 
     @Override
@@ -97,7 +97,7 @@ public class ExpressionContext {
 
         private int maxStringLength;
         private DateTimeSettings dateTimeSettings;
-        private StateFetcher stateFetcher;
+        private StateProvider stateProvider;
 
         private Builder() {
         }
@@ -105,7 +105,7 @@ public class ExpressionContext {
         private Builder(final ExpressionContext expressionContext) {
             this.maxStringLength = expressionContext.maxStringLength;
             this.dateTimeSettings = expressionContext.dateTimeSettings;
-            this.stateFetcher = expressionContext.stateFetcher;
+            this.stateProvider = expressionContext.stateProvider;
         }
 
         public Builder maxStringLength(final int maxStringLength) {
@@ -118,13 +118,13 @@ public class ExpressionContext {
             return this;
         }
 
-        public Builder stateFetcher(final StateFetcher stateFetcher) {
-            this.stateFetcher = stateFetcher;
+        public Builder stateProvider(final StateProvider stateProvider) {
+            this.stateProvider = stateProvider;
             return this;
         }
 
         public ExpressionContext build() {
-            return new ExpressionContext(maxStringLength, dateTimeSettings, stateFetcher);
+            return new ExpressionContext(maxStringLength, dateTimeSettings, stateProvider);
         }
     }
 }